requirements.txt
cffi==2.0.0
Cython==3.2.3
ffmpeg-python==0.2.0
future==1.0.0
madmom @ git+https://github.com/CPJKU/madmom@27f032e8947204902c675e5e341a3faf5dc86dae
mido==1.3.3
numpy==2.3.5
packaging==25.0
pycparser==2.23
pytest-qt==4.5.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-vlc==3.0.21203
scipy==1.16.3
shiboken6==6.10.0
sounddevice==0.5.3
soundfile==0.13.1
